Уберет моки 1
