+++ src/presenter/single-waypoint-presenter.js
@@ -2,7 +2,7 @@ import { render } from '../framework/render.js';
 import EventEditView from '../view/event-edit-view.js';
 import EventView from '../view/event-view.js';
 import { replace, remove } from '../framework/render.js';
-import { Mode, UpdateType, UserAction } from '../mock/const.js';
+import { Mode, UpdateType, UserAction } from '../util/const.js';
 
 export default class SingleWaypointPresenter {
   #eventListContainer = null;
