+++ src/presenter/waypoints-presenter.js
@@ -2,10 +2,10 @@ import { remove, render } from '../framework/render.js';
 import EventListView from '../view/event-list-view.js';
 import EmptyListMessage from '../view/event-list-empty-view.js';
 import SingleWaypointPresenter from './single-waypoint-presenter.js';
-import { sortPointByPrice, sortPointByTime } from '../mock/waypoints.js';
+import { sortPointByPrice, sortPointByTime } from '../util/waypoints.js';
 import SortView from '../view/sort-view.js';
-import { SortType, UpdateType, UserAction, FilterType } from '../mock/const.js';
-import { filter } from '../mock/filter.js';
+import { SortType, UpdateType, UserAction, FilterType } from '../util/const.js';
+import { filter } from '../util/filter.js';
 import NewWaypointPresenter from './new-waypoint-presenter.js';
 import WaypointLoadingListView from '../view/waypoints-loading-list-view.js';
 import UiBlocker from '../framework/ui-blocker/ui-blocker.js';
