+++ src/util/const.js
@@ -19,14 +19,6 @@ export const WAYPOINT_TYPES = [
   'Restaurant',
 ];
 
-export const OFFERS = [
-  'Upgrade to a business class',
-  'Add luggage',
-  'Switch to comfort',
-  'Rent a car',
-  'Order Uber',
-  'Add breakfast',
-];
 
 export const Mode = {
   DEFAULT: 'DEFAULT',
@@ -59,17 +51,4 @@ export const UpdateType = {
   INIT: 'INIT',
 };
 
-export const DATE_FROM = [
-  '2019-06-10T22:55:56.845Z',
-  '2019-07-10T21:30:56.845Z',
-  '2019-07-10T19:15:56.845Z',
-  '2019-07-10T12:10:31.845Z',
-];
-
-export const DATE_TO = [
-  '2019-07-11T12:10:13.375Z',
-  '2019-07-11T13:05:07.375Z',
-  '2019-07-11T10:33:13.375Z',
-  '2019-07-11T14:44:22.375Z',
-];
