+++ src/util/filter.js
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs';
 import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
-import { FilterType } from './const.js';
+import { FilterType } from '../util/const.js';
 
 dayjs.extend(isSameOrAfter);
 dayjs.extend(isSameOrBefore);
