+++ src/view/event-list-empty-view.js
@@ -1,5 +1,6 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { FilterType } from '../mock/const.js';
+import { FilterType } from '../util/const.js';
+
 
 const NoWaypointTextType = {
   [FilterType.EVERYTHING]: 'Click New Event to create your first point',
