+++ src/view/sort-view.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { SortType } from '../mock/const.js';
+import { SortType } from '../util/const.js';
 
 function createSortTemplate(currentSortType) {
   return `
